se agrega boton para cancelar a la hora de editar en la pagina de administracion. Se agrega confirmacion a ala hora de eliminar en adminostracion. Se corrige toda la pagina de mis mascotas y ahora se puede agregar mascotas y eliminarlas. Se corrige error de key en tablas. Se corrigio contexts

+++ VeterinariaHuellitas/src/components/tablas/TablaUsuarios.jsx
@@ -21,13 +21,22 @@ const TablaUsuarios = () => {
     };
 
     const handleDelete = (id) => {
-        eliminarUsuario(id);
         Swal.fire({
-            position: "center",
-            icon: "success",
-            title: "Usuario eliminado",
+          title: '¿Estás seguro?',
+          text: 'Esta acción eliminará el usuario.',
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#3085d6',
+          cancelButtonColor: '#d33',
+          confirmButtonText: 'Sí, eliminar',
+          cancelButtonText: 'Cancelar',
+        }).then(async (result) => {
+          if (result.isConfirmed) {
+            await eliminarUsuario(id);
+            Swal.fire('Usuario eliminado', '', 'success');
+          }
         });
-    };
+      };
 
     return (
         <>
